SAA-1412 offender merge changes for allocations.

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsactivitiesmanagementapi/common/OffenderMergeDetails.kt
@@ -1,3 +1,3 @@
 package uk.gov.justice.digital.hmpps.hmppsactivitiesmanagementapi.common
 
-data class OffenderMergeDetails(val prisonCode: String, val newNumber: String, val oldNumber: String, val newBookingId: Int? = null)
+data class OffenderMergeDetails(val prisonCode: String, val newNumber: String, val oldNumber: String, val newBookingId: Long? = null)
