+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsactivitiesmanagementapi/service/events/handlers/OffenderMergedEventHandler.kt
@@ -53,7 +53,7 @@ class OffenderMergedEventHandler(
                 prisonCode = prisonId,
                 oldNumber = oldNumber,
                 newNumber = newNumber,
-                newBookingId = prisoner.bookingId?.toInt(),
+                newBookingId = prisoner.bookingId?.toLong(),
               ),
             )
           }
